Move git code to separate submodule
